More cleanup and simplification
